pyproject.toml
[tool.poetry]
name = "threatfusion"
version = "0.1.0"
description = "Automated Threat Intelligence Aggregator"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
pydantic = "^2.5.3"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
rich = "^13.7.0"
jinja2 = "^3.1.2"
weasyprint = "^60.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.12.1"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[tool.poetry.scripts]
threatfusion = "src.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
